🚚 Move FileKit code into filekit-core
