+++ filekit-core/src/appleMain/kotlin/io/github/vinceglb/filekit/FileKit.apple.kt
@@ -1,0 +1,3 @@
+package io.github.vinceglb.filekit
+
+public actual object FileKit
