+++ filekit-core/src/commonMain/kotlin/io/github/vinceglb/filekit/FileKit.kt
@@ -1,0 +1,3 @@
+package io.github.vinceglb.filekit
+
+public expect object FileKit
