+++ filekit-core/src/jsMain/kotlin/io/github/vinceglb/filekit/FileKit.js.kt
@@ -1,0 +1,3 @@
+package io.github.vinceglb.filekit
+
+public actual object FileKit
