+++ filekit-core/src/jvmMain/kotlin/io/github/vinceglb/filekit/FileKit.jvm.kt
@@ -1,0 +1,3 @@
+package io.github.vinceglb.filekit
+
+public actual object FileKit
