+++ filekit-core/src/wasmJsMain/kotlin/io/github/vinceglb/filekit/FileKit.wasmJs.kt
@@ -1,0 +1,3 @@
+package io.github.vinceglb.filekit
+
+public actual object FileKit
