+++ filekit-dialog/src/commonMain/kotlin/io/github/vinceglb/filekit/dialog/FileKit.kt
@@ -1,8 +1,7 @@
 package io.github.vinceglb.filekit.dialog
 
+import io.github.vinceglb.filekit.FileKit
 import io.github.vinceglb.filekit.PlatformFile
-
-public expect object FileKit
 
 public expect suspend fun <Out> FileKit.pickFile(
     type: PickerType = PickerType.File(),
