+++ filekit-dialog/src/nonWebMain/kotlin/io/github/vinceglb/filekit/dialog/FileKit.nonWeb.kt
@@ -1,5 +1,6 @@
 package io.github.vinceglb.filekit.dialog
 
+import io.github.vinceglb.filekit.FileKit
 import io.github.vinceglb.filekit.PlatformFile
 
 public expect suspend fun FileKit.pickDirectory(
