+++ samples/sample-core/composeApp/src/androidMain/kotlin/io/github/vinceglb/sample/core/compose/MainActivity.kt
@@ -3,7 +3,7 @@ package io.github.vinceglb.sample.core.compose
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import io.github.vinceglb.filekit.dialog.FileKit
+import io.github.vinceglb.filekit.FileKit
 
 class MainActivity : ComponentActivity() {
 	override fun onCreate(savedInstanceState: Bundle?) {
